Ported the main menu to python code
I was able to break out almost all the object declaration from the .kv files
that were controlling the look of the main menu and put them into a .py file.
This will allow us to easily update these elements in the future if we see a
need for it (for things such as dark mode or font resizing.