Changed language selection button
Changed langauge selection button to show default language defined
within the config file. On selection, button should update with the
newly selected language and replace the location in the list with
the previously selected language.

Currently bugged, language does not update button correctly when
selected, but the language does change in the app. This is purely
a visual bug.